Correct translation content from reusables too (#37722)
