+++ lib/get-data.js
@@ -6,6 +6,7 @@ import matter from 'gray-matter'
 import { merge, get } from 'lodash-es'
 
 import languages from './languages.js'
+import { correctTranslatedContentStrings } from './correct-translation-content.js'
 
 // If you run `export DEBUG_JIT_DATA_READS=true` in your terminal,
 // next time it will mention every file it reads from disk.
@@ -175,7 +176,33 @@ function getDataByDir(dottedPath, dir, englishRoot) {
     fullPath.push(...split)
     fullPath.push(`${nakedname}.md`)
     const markdown = getMarkdownContent(dir, fullPath.join(path.sep), englishRoot)
-    return matter(markdown).content
+    let { content } = matter(markdown)
+    if (dir !== englishRoot) {
+      // If we're reading a translation, we need to replace the possible
+      // corruptions. For example `[AUTOTITLE"을](/foo/bar)`.
+      // To do this we'll need the English equivalent
+      let englishContent = content
+      try {
+        englishContent = getMarkdownContent(englishRoot, fullPath.join(path.sep), englishRoot)
+      } catch (error) {
+        // In some real but rare cases a reusable doesn't exist in English.
+        // At all.
+        // This can happen when the translation is really out of date.
+        // You might have an old `docs-internal.locale/content/**/*.md`
+        // file that mentions `{% data reusables.foo.bar %}`. And it's
+        // working fine, except none of that exists in English.
+        // If this is the case, we still want to executed the
+        // correctTranslatedContentStrings() function, but we can't
+        // genuinely give it the English equivalent content, which it
+        // sometimes uses to correct some Liquid tags. At least other
+        // good corrections might happen.
+        if (error.code !== 'ENOENT') {
+          throw error
+        }
+      }
+      content = correctTranslatedContentStrings(content, englishContent)
+    }
+    return content
   }
 
   // E.g. {% data ui.pages.foo.bar %}
