+++ middleware/contextualizers/glossaries.js
@@ -1,7 +1,7 @@
 import { getDataByLanguage } from '../../lib/get-data.js'
 import liquid from '../../lib/render-content/liquid.js'
 import { executeWithFallback } from '../../lib/render-with-fallback.js'
-import { correctTranslatedContentStrings } from '../../lib/page-data.js'
+import { correctTranslatedContentStrings } from '../../lib/correct-translation-content.js'
 
 export default async function glossaries(req, res, next) {
   if (!req.pagePath.endsWith('get-started/quickstart/github-glossary')) return next()
