+++ tests/fixtures/translations/ja-jp/data/reusables/gated-features/more-info.md
@@ -1,0 +1,1 @@
+{% ifversion fpt or ghec %}詳細については, see ["AUTOTITLE](/get-started/quickstart/hello-world)."{% endif %}
